Adds FindAsync methods to Crud Repositories
Adds `FindAsync` methods to both the base Crud repository and the DTO Crud repository interfaces and implementations.

These methods allow retrieval of a single entity or DTO based on a filter expression, with options to throw an exception if multiple entities are found, or to ignore "deleted" entities in auditable tables.

This addition enhances the data access layer by providing more flexible and controlled retrieval options.